Option-Pricing-Model
Options pricing models are commonly leveraged by quantitative traders and researchers to calculate the theoretical value of an option. Understanding the theoretical value of an option can allow investors to adjust their trading strategies in a way that aligns with the set of options that would reap the most profits. Since there isn't a single definitive way to calculate the theoretical value of an option, many different techniques have arisen. For this project, you can implement one of these pricing models using the programming language of your choice.